lot name set to computed, stored, editable
Instead of setting its name as a default lambda expression, it's name is computed the first time the product is assigned.  This way we can use the product-specific sequence if it's set, and if it's not set we use the default sequence.

We only set name if it is empty a) to not overwrite a user edit and b) to ensure the sequence is only incremented one time, like when it was the default lambda.  This best mirrors existing behaviour, but means if you change product the s/n will not auto-update to the new sequence.